tests/submission/Infraware/Forms/contents/Forms/progress_value.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>progress_value</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
		 <p>Progress: <progress max=100 id='progress_id'></progress></p>
	</form>

	<script>
		
		var progress = document.getElementById("progress_id");
		progress.value = 50;

		if (typeof(progress.value) == "number") {
			test(function() {
				assert_equals(progress.value, 50, "value attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("value attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
